requirements.txt
flask
flask-cors
requests
beautifulsoup4
yfinance
scikit-learn
numpy
lxml
statsmodels
python-dotenv
pymysql>=1.1.0
cryptography>=41.0.0
bcrypt>=4.1.2
pywebpush>=2.0.3
gunicorn
waitress